# cmd/swagger/docs/swagger.yaml
basePath: /api
definitions:
...
  dto.AddGameReq:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  dto.AddPlayerReq:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  dto.Claim:
    properties:
      player_id:
        type: string
      user_id:
        type: string
    type: object
  dto.CreateGroupReq:
    properties:
      name:
        type: string
      player_name:
        type: string
    required:
    - name
    - player_name
    type: object
...
  dto.Game:
    properties:
      bgg_id:
        type: integer
      id:
        type: string
      name:
        type: string
      original:
        type: string
    type: object
...
  dto.Group:
    properties:
      id:
        type: string
      name:
        type: string
      owner:
        type: string
    type: object
  dto.Invite:
    properties:
      expires:
        type: string
      id:
        type: string
      revoked:
        type: boolean
    type: object
  dto.InvitePreview:
    properties:
      group_id:
        type: string
      name:
        type: string
    type: object
  dto.InviteResp:
    properties:
      expires:
        type: string
      id:
        type: string
      url:
        type: string
    type: object
  dto.JoinReq:
    properties:
      token:
        type: string
    required:
    - token
    type: object
...
  dto.ManageReq:
    properties:
      action:
        type: string
      target:
        type: string
      value:
        type: string
    required:
    - action
    type: object
  dto.Member:
    properties:
      email:
        type: string
      user_id:
        type: string
    type: object
  dto.Notification:
    properties:
      body:
        type: string
      created:
        type: string
      group_id:
        type: string
      id:
        type: string
      kind:
        type: string
      link:
        type: string
      read_at:
        type: string
      title:
        type: string
    type: object
  dto.NotificationPage:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.Notification'
        type: array
      unread:
        type: integer
    type: object
...
  dto.Player:
    properties:
      account:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  dto.Snapshot:
    properties:
      claims:
        items:
          $ref: '#/definitions/dto.Claim'
        type: array
      games:
        items:
          $ref: '#/definitions/dto.Game'
        type: array
      group:
        $ref: '#/definitions/dto.Group'
      locations:
        items:
          type: string
        type: array
      members:
        items:
          $ref: '#/definitions/dto.Member'
        type: array
      players:
        items:
          $ref: '#/definitions/dto.Player'
        type: array
    type: object
...
  ginutils.Ret-any:
    properties:
      code:
        type: integer
      data: {}
      message: {}
    type: object
...
  ginutils.Ret-array_dto_Group:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/dto.Group'
        type: array
      message: {}
    type: object
  ginutils.Ret-array_dto_Invite:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/dto.Invite'
        type: array
      message: {}
    type: object
...
  ginutils.Ret-dto_Game:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/dto.Game'
      message: {}
    type: object
  ginutils.Ret-dto_Group:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/dto.Group'
      message: {}
    type: object
  ginutils.Ret-dto_InvitePreview:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/dto.InvitePreview'
      message: {}
    type: object
  ginutils.Ret-dto_InviteResp:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/dto.InviteResp'
      message: {}
    type: object
...
  ginutils.Ret-dto_NotificationPage:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/dto.NotificationPage'
      message: {}
    type: object
...
  ginutils.Ret-dto_Player:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/dto.Player'
      message: {}
    type: object
...
  ginutils.Ret-dto_Snapshot:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/dto.Snapshot'
      message: {}
    type: object
...
  ginutils.Ret-string:
    properties:
      code:
        type: integer
      data:
        type: string
      message: {}
    type: object
info:
  contact: {}
  description: |-
    熟人小组对局日记的接口文档；路径相对 /api 挂载点。
    除图片读取外，所有接口返回 ginutils.Ret 包络：{code, data, message}；code 为稳定业务码，失败时 HTTP 状态同时表达语义。
  title: 又一局 API
  version: "1.0"
paths:
...
  /v1/auth/group-invite:
    post:
      consumes:
      - application/json
      description: 无需登录，凭有效邀请仅返回小组 ID 与名称，不展示成员或记录
      parameters:
      - description: 邀请令牌
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.JoinReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ginutils.Ret-dto_InvitePreview'
      summary: 查看小组邀请
      tags:
      - Auth
...
  /v1/groups:
    get:
      description: 返回当前账号加入的小组
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ginutils.Ret-array_dto_Group'
      security:
      - SessionCookie: []
      summary: 获取我的小组
      tags:
      - Group
    post:
      consumes:
      - application/json
      description: 创建小组与创建者自己的玩家档案，创建者成为组主与首位成员
      parameters:
      - description: 创建小组请求体
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateGroupReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ginutils.Ret-dto_Group'
      security:
      - SessionCookie: []
      summary: 创建小组
      tags:
      - Group
  /v1/groups/{group}:
    get:
      description: 返回小组成员、玩家、桌游、待确认关联与最近地点
      parameters:
      - description: 小组 ID
        in: path
        name: group
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ginutils.Ret-dto_Snapshot'
      security:
      - SessionCookie: []
      summary: 获取小组详情
      tags:
      - Group
  /v1/groups/{group}/bgg/search:
    get:
      description: 外部桌游检索；授权未配置时返回明确不可用提示
      parameters:
      - description: 小组 ID
        in: path
        name: group
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ginutils.Ret-any'
      security:
      - SessionCookie: []
      summary: 搜索外部桌游
      tags:
      - Game
  /v1/groups/{group}/export:
    get:
      description: 仅组主可下载包含 JSON、CSV 与已关联照片的 ZIP
      parameters:
      - description: 小组 ID
        in: path
        name: group
        required: true
        type: string
      produces:
      - application/zip
      responses:
        "200":
          description: OK
          schema:
            type: file
      security:
      - SessionCookie: []
      summary: 导出小组备份
      tags:
      - Group
  /v1/groups/{group}/games:
    post:
      consumes:
      - application/json
      description: 在小组内添加桌游，同组同名复用已有条目
      parameters:
      - description: 小组 ID
        in: path
        name: group
        required: true
        type: string
      - description: 添加桌游请求体
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AddGameReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ginutils.Ret-dto_Game'
      security:
      - SessionCookie: []
      summary: 添加桌游
      tags:
      - Game
  /v1/groups/{group}/invites:
    get:
      description: 组主查看本组邀请记录，不返回明文令牌
      parameters:
      - description: 小组 ID
        in: path
        name: group
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ginutils.Ret-array_dto_Invite'
      security:
      - SessionCookie: []
      summary: 获取小组邀请
      tags:
      - Group
    post:
      description: 组主生成七天有效的小组邀请，明文令牌仅返回一次
      parameters:
      - description: 小组 ID
        in: path
        name: group
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ginutils.Ret-dto_InviteResp'
      security:
      - SessionCookie: []
      summary: 生成小组邀请
      tags:
      - Group
  /v1/groups/{group}/manage:
    post:
      consumes:
      - application/json
      description: 组主或本人执行改名、转让、移除、撤销邀请、关联与审核玩家
      parameters:
      - description: 小组 ID
        in: path
        name: group
        required: true
        type: string
      - description: 小组操作请求体
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ManageReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ginutils.Ret-any'
      security:
      - SessionCookie: []
      summary: 修改小组与成员
      tags:
      - Group
...
  /v1/groups/{group}/players:
    post:
      consumes:
      - application/json
      description: 在小组内添加昵称玩家，重名会被拒绝
      parameters:
      - description: 小组 ID
        in: path
        name: group
        required: true
        type: string
      - description: 添加玩家请求体
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AddPlayerReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ginutils.Ret-dto_Player'
      security:
      - SessionCookie: []
      summary: 添加玩家
      tags:
      - Group
...
  /v1/join:
    post:
      consumes:
      - application/json
      description: 使用小组邀请链接中的令牌加入小组
      parameters:
      - description: 加入小组请求体
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.JoinReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ginutils.Ret-string'
      security:
      - SessionCookie: []
      summary: 加入小组
      tags:
      - Group
...
  /v1/notifications:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ginutils.Ret-dto_NotificationPage'
      security:
      - SessionCookie: []
      summary: 获取站内通知
      tags:
      - Notification
  /v1/notifications/{id}/read:
    post:
      parameters:
      - description: 通知 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ginutils.Ret-any'
      security:
      - SessionCookie: []
      summary: 标记通知已读
      tags:
      - Notification
...
securityDefinitions:
  SessionCookie:
    in: cookie
    name: omr_session
    type: apiKey
swagger: "2.0"